Ferdig med datepicker, litt småfiks gjenstår
